refactor: separate resource and registry bridging for clarity
